Spring cleaning. Prepare for summer! Fixed lots of bugs too: v_camroll is now working again. Counter-Strike is finally getting somewhere again. Counter-Strike has pseudo spray-patterns now! They're all the same. func_button now can be programmed to use Sound-Shaders. The menu handles the vid_conautoscale command autonomously. Set r_autoscale to 0 if you don't like it. The menu has an updater, requires a not-yet-pushed build of FTE. Don't use it yet. Moved a lot of globals into pSeat's structs to fix splitscreen with them. Made explosion decals be handled client-side! Added trigger_look, env_sun and light_environment entities.

+++ src/client/valve/cmds.c
@@ -14,7 +14,8 @@
  * OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
 
-float Game_ConsoleCommand(void)
+int
+Game_ConsoleCommand(void)
 {
 	switch(argv(0)) {
 		default:
